HPCC-35881 Bump vcpkg to latest version

+++ ecl/hqlcpp/hqlhtcpp.cpp
@@ -4021,7 +4021,7 @@ unsigned HqlCppTranslator::buildRtlIfBlockField(StringBuffer & instanceName, IHq
     BuildCtx declarectx(*code, declareAtom);
 
     //First generate a pseudo type entry for an ifblock.
-    unsigned fieldType = type_ifblock|RFTMunknownsize;
+    unsigned fieldType = (unsigned)type_ifblock|RFTMunknownsize;
     {
         StringBuffer childTypeName;
         unsigned childType = buildRtlRecordFields(childTypeName, ifblock->queryChild(1), rowRecord, rowTypeName);
